# Conflicts:
#	modules/dfs/dfsmanager.py

+++ modules/dfs/dfsmanager.py
@@ -106,8 +106,16 @@ class DFSManager:
         self._filewriter.write_to_replica(filename, part, total, data)
 
     def dump_replica(self, filename):
-        ## remove from disk
-        self._filewriter.remove(filename)
+        file = self._fs.get_file(filename)
+        if not file:
+            print("Invalid name")
+            return
+
+        # Check that you are a replica before removing file from disk
+        file_replicas = file["replicas"]
+        if self._id in file_replicas:
+            self._filewriter.remove(filename)
+
         ## remove from _fs
         self._fs.delete_file(filename)
 
@@ -146,10 +154,24 @@ class DFSManager:
 
     def delete_file(self, filename):
         ## remove from disk (if present)
-        self._filewriter.remove(filename)
+        ## remove from _fs
+        ## tell network to tell replicas
+        file = self._fs.get_file(filename)
+        if not file:
+            print("Invalid name")
+            return
+
+        # Check that you are a replica before removing file from disk
+        file_replicas = file["replicas"]
+        if self._id in file_replicas:
+            self._filewriter.remove(filename)
+
         ## remove from _fs
         self._fs.delete_file(filename)
+
         ## tell network to tell replicas
+        self._network.delete_file(filename)
+
         print("Deleted %s" % (filename))
 
     def set_priority(self, filename, priority):
